# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FoodImageUpload {
  image: Blob;
}

export interface FoodModelInfo {
  name: string;
  algorithm: string;
  version: string;
  trainedSamples: number;
  features: string[];
  classes: string[];
  trainingNote: string;
}

export type FoodPipelineStageStatus = typeof FoodPipelineStageStatus[keyof typeof FoodPipelineStageStatus];


export const FoodPipelineStageStatus = {
  ready: 'ready',
  optional: 'optional',
} as const;

export interface FoodPipelineStage {
  id: string;
  label: string;
  description: string;
  status: FoodPipelineStageStatus;
}

export interface FoodPipelineInfo {
  stages: FoodPipelineStage[];
  ocrEngine: string;
  visionLibrary: string;
  classifier: string;
}

export interface FoodExtractedField {
  value: string;
  confidence: number;
  source: string;
}

export interface FoodExtractedInfo {
  productName: FoodExtractedField;
  brand: FoodExtractedField;
  manufacturingDate: FoodExtractedField;
  expiryDate: FoodExtractedField;
  bestBefore: FoodExtractedField;
  batchNumber: FoodExtractedField;
  ingredients: FoodExtractedField;
  mrp: FoodExtractedField;
}

export interface FoodProbability {
  label: string;
  probability: number;
}

export type FoodRiskIndicatorSeverity = typeof FoodRiskIndicatorSeverity[keyof typeof FoodRiskIndicatorSeverity];


export const FoodRiskIndicatorSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface FoodRiskIndicator {
  label: string;
  severity: FoodRiskIndicatorSeverity;
  detail: string;
}

export type FoodAnalysisExpiryStatus = typeof FoodAnalysisExpiryStatus[keyof typeof FoodAnalysisExpiryStatus];


export const FoodAnalysisExpiryStatus = {
  expired: 'expired',
  not_expired: 'not_expired',
  unavailable: 'unavailable',
} as const;

export type FoodAnalysisRiskLevel = typeof FoodAnalysisRiskLevel[keyof typeof FoodAnalysisRiskLevel];


export const FoodAnalysisRiskLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface FoodAnalysis {
  analysisId: string;
  filename: string;
  analyzedAt: string;
  imagePreview: string;
  expiryStatus: FoodAnalysisExpiryStatus;
  expiryMessage: string;
  expiryConfidence: number;
  foodCategory: string;
  categoryConfidence: number;
  riskLevel: FoodAnalysisRiskLevel;
  riskConfidence: number;
  riskProbabilities: FoodProbability[];
  riskIndicators: FoodRiskIndicator[];
  extracted: FoodExtractedInfo;
  rawOcrText: string;
  preprocessing: string[];
  model: FoodModelInfo;
  disclaimer: string;
}

export interface ErrorResponse {
  error: string;
}