# lib/api-zod/src/generated/types/foodAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FoodAnalysisExpiryStatus } from './foodAnalysisExpiryStatus';
import type { FoodAnalysisRiskLevel } from './foodAnalysisRiskLevel';
import type { FoodExtractedInfo } from './foodExtractedInfo';
import type { FoodModelInfo } from './foodModelInfo';
import type { FoodProbability } from './foodProbability';
import type { FoodRiskIndicator } from './foodRiskIndicator';

export interface FoodAnalysis {
  analysisId: string;
  filename: string;
  analyzedAt: string;
  imagePreview: string;
  expiryStatus: FoodAnalysisExpiryStatus;
  expiryMessage: string;
  expiryConfidence: number;
  foodCategory: string;
  categoryConfidence: number;
  riskLevel: FoodAnalysisRiskLevel;
  riskConfidence: number;
  riskProbabilities: FoodProbability[];
  riskIndicators: FoodRiskIndicator[];
  extracted: FoodExtractedInfo;
  rawOcrText: string;
  preprocessing: string[];
  model: FoodModelInfo;
  disclaimer: string;
}