# lib/api-zod/src/generated/types/foodExtractedField.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FoodExtractedField {
  value: string;
  confidence: number;
  source: string;
}