# lib/api-zod/src/generated/types/foodExtractedInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FoodExtractedField } from './foodExtractedField';

export interface FoodExtractedInfo {
  productName: FoodExtractedField;
  brand: FoodExtractedField;
  manufacturingDate: FoodExtractedField;
  expiryDate: FoodExtractedField;
  bestBefore: FoodExtractedField;
  batchNumber: FoodExtractedField;
  ingredients: FoodExtractedField;
  mrp: FoodExtractedField;
}