# lib/api-zod/src/generated/types/foodModelInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FoodModelInfo {
  name: string;
  algorithm: string;
  version: string;
  trainedSamples: number;
  features: string[];
  classes: string[];
  trainingNote: string;
}