# lib/api-zod/src/generated/types/foodRiskIndicator.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FoodRiskIndicatorSeverity } from './foodRiskIndicatorSeverity';

export interface FoodRiskIndicator {
  label: string;
  severity: FoodRiskIndicatorSeverity;
  detail: string;
}